requirements.txt
Django==3.1.4
newsapi-python==0.2.6


